requirements.txt
pytest==6.2.2
mypy==0.812
pylint==2.7.4
pycairo==1.20.1
